Uniquify the automatically assigned object-ids for named script entries.
Problem:
Object-ids for named script entries used to be assigned from 4716 upwards.
This leads to projects which all share the first object-ids. The pen then
cannot detect when a book/product was not correctly activated and the
behavior (sounds etc) does not match the book.

Solution:
Use different object-ids for different projects as far as possible.
Different projects always use a different product id, so the starting
point of automatically assigned object ids is derived from the
product id.

The assignemnt is scattered across the range of valid object-ids in such
a way that the average distance (in terms of object-ids) between projects
is always maximized independent of the number of projects involved.