Implement reading and setting the language header field
The syntax for the YAML file is, for example:
```
gme-lang: FRENCH
```
The program does not check whether the language string makes sense and
will use it as-is, captialized and all. This fixes #105.

Also do not confuse this with the `language:` field, which specifies the
language to be used for text-to-speech.