More nix experiments
`nix/static` is a plain `nixpkgs` setup, and almost builds a static
version of the binary, if it were not for
https://github.com/nh2/static-haskell-nix/issues/14

`nix/` is using the tooling at https://github.com/input-output-hk/iohk-nix
as advised by @angerman, but it currenty fails to download or build
`/nix/store/j6m77ilz08bsdhs8cba9i0mhs4lw0vak-ghc-8.4.4.drv` because some
patches do not apply. Or something.