A test suite
this is long due: The beginning of a test suite for `tttool`, starting
with simple “golden tests” that assert that certain tool interactions
are unchanged.

Also added, a `cabal.project.freeze` to ensure the same version of
`yaml` is used (else the output differs). CI tests ensure that this file
is consistent with the `nix` view.