libtiptoi: length and verbose output (#220)
* libtiptoi: print the sizes of the file

I was wondering but the actual problem was and decided to make the
output a bit more verbose.

Note how the data types are different, i.e. uint32_t and long int.
Someone ought to use more appropriate types. Maybe size_t.

* libtiptoi: fixed "length" typo

By using the correct spelling, the code is a bit easier on the eye.

The change is a very mechanical search and replace.