Audio table decoding: Detect the audio table copy more liberally
Most or all Ravensburger GMEs have a duplicated Audio table. Most of
them have an identical copy, but it seems not all. For those where the
table is 80% equal, treat it like a copy (and throw away the extra
information for now).

It remains future work to make sense of this extra table.