lint: verify that each line contains only one jump
As discussed in #231 and #238, TipToi pens handle a jump after all other
actions (even if they are earlier in the list). If there are multiple
jumps, only the last has an effect. This lint check verifies that a GME
only contains one jump, and that the jump is the last action. This check
is true for all GMEs published by Ravensburger.